refactor: chat service interface 분리
