+++ src/main/java/com/tasksprints/auction/api/bid/BidController.java
@@ -5,7 +5,7 @@ import com.tasksprints.auction.common.response.ApiResult;
 import com.tasksprints.auction.domain.bid.dto.BidRequest;
 import com.tasksprints.auction.domain.bid.dto.BidResponse;
 import com.tasksprints.auction.domain.bid.service.BidService;
-import com.tasksprints.auction.domain.socket.service.ChatService;
+import com.tasksprints.auction.domain.socket.service.ChatServiceImpl;
 import com.tasksprints.auction.domain.user.dto.response.UserDetailResponse;
 import com.tasksprints.auction.domain.user.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -24,7 +24,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @RequestMapping("/api/v1/bid")
 public class BidController {
     private final BidService bidService;
-    private final ChatService chatService;
+    private final ChatServiceImpl chatService;
     private final UserService userService;
     private final SimpMessageSendingOperations simpMessageSendingOperations;
